Generalize code to work with meshes that have fewer than 22000000 cells
In order to avoid writing more than 2 GB at a time, we previously split up
the writing of fields into chunks that were sized at most 22000000 (or in
the case of weightsOnEdge, 11000000) in the outer dimension. However, this
logic fails when the mesh to be written has fewer than 22000000 cells.

This commit generalizes the chunking logic so that we write chunks that are
at most min(nCells,22000000) at a time.