Correct writing of edge and vertex coordinate fields for large meshes
The {lat,lon,x,y,z}x{Edge,Vertex} fields can exceed 2^31 bytes in size
for large meshes (e.g., the 2-km mesh with 147M cells), and writing this
much data in one call seems to be a problem for the Parallel-NetCDF
library. This commit writes these fields in the same way as is done
for other large indexing fields: by writing chunks at a time to keep
the size of individual writes below 2 GB.

This commit also changes a few index variables used to track edges used
in the computation of weightsOnEdge from 'int' to 'size_t' types.