Fix bug in computation of angleEdge
The previous method for computing angleEdge assumed that local variables 'vert1' and
'vert2' gave the indices of verticesOnEdge[i][0] and verticesOnEdge[i][1], respectively.
However, these two indices are taken from an unordered set, and hence may be in the opposite
order.

We now compute the angleEdge field using a more accurate form based on (but not exactly
the same as) the code used in grid_rotate, properly accounting for the possible transposition
of the order of vert1 and vert2.