Add initial version of permute_mesh utility
The permute_mesh utility may be used to permute the order of cells, edges,
and vertices in a mesh or atmosphere static file. At present, the only
permutation that is supported is one which guarantees that each MPI task
owns a contiguous range of cells, edges, and vertices in the permuted output
file, when the decomposition of the input mesh is supplied with
a graph.info.part.X file.

This utility is helpful, e.g., to arrange a mesh for optimal I/O, so that
each MPI task can independently read and write contiguous ranges of bytes
from files in parallel.