Add initial support for remapping 1-d (2-d spatial) fields
This commit adds code to the NCField class as well as to the Remapper classes
to handle interpolation of 2-d horizontal fields with no time dimension
(e.g., 'ter'), which makes them 1-d arrays.

Also included in this commit are changes to allow horizontal remapping weights
to be computed without computing vertical remapping weights. This can be
useful in situations where we don't have vertical grid information but need to
remap, e.g., static fields. In order to generate horizontal weights only,
the nVertLevelsSrc and nVertLevelsDst arguments to the computeWeights{Cell,Edge}
routines must both be set to 0; in this case, the levelsSrc and levelsDst
arguments can be NULL.