Add command-line option --use-reconstruct-winds
Currently, this option doesn't actually do anything, but will eventually
allow for the use of the uReconstructZonal and uReconstructMeridional
winds to be used instead of the 'u' field.

This commit also includes some basic checks that we can read all other
input files to hopefully catch cases where options are being used as
filenames.